+++ docs/ChangeLog.md
@@ -5,7 +5,8 @@ UNRELEASED VERSION
 ----
 
 - Adds {RPicSim::Memory#read_bytes} and {RPicSim::Memory#write_bytes}.
-- Fixed a bug in the RSpec 3.x suppoort that was causing the informative error messages when an example fails to not work.
+- Adds {RPicSim::Sim#stack_memory}.
+- Fixes a bug in the RSpec 3.x suppoort that was causing the informative error messages when an example fails to not work.
 
 0.3.0
 ----